+++ Flameberry/src/Core/Timer.h
@@ -6,10 +6,10 @@ namespace Flameberry {
     class ScopedTimer
     {
     public:
-        ScopedTimer(float* durationVar);
+        ScopedTimer(double* durationVar);
         ~ScopedTimer();
     private:
-        float* m_DurationVar;
+        double* m_DurationVar;
         bool m_Started = false;
         std::__1::chrono::steady_clock::time_point m_Start;
     };
